+++ sda-commons-server-jackson/src/main/java/org/sda/commons/server/jackson/hal/HalLinkMethodInvocationException.java
@@ -1,6 +1,14 @@
 package org.sda.commons.server.jackson.hal;
 
-public class HalLinkMethodInvocationException extends RuntimeException {
+/**
+ * @deprecated this package has been created by mistake. The {@code
+ *     HalLinkMethodInvocationException} moved to {@link
+ *     org.sdase.commons.server.jackson.hal.HalLinkMethodInvocationException}, please update the
+ *     imports.
+ */
+@Deprecated
+public class HalLinkMethodInvocationException
+    extends org.sdase.commons.server.jackson.hal.HalLinkMethodInvocationException {
 
   public HalLinkMethodInvocationException(String message) {
     super(message);
