+++ sda-commons-server-jackson/src/main/java/org/sda/commons/server/jackson/hal/LinkResult.java
@@ -1,13 +1,16 @@
 package org.sda.commons.server.jackson.hal;
 
 import io.openapitools.jackson.dataformat.hal.HALLink;
-import io.openapitools.jackson.dataformat.hal.HALLink.Builder;
 import java.net.URI;
 
-/** Wrapper class to provide the processed Link as {@linkplain URI} or {@linkplain HALLink} */
-public final class LinkResult {
-  private final URI uri;
-  private final HALLink halLink;
+/**
+ * Wrapper class to provide the processed Link as {@linkplain URI} or {@linkplain HALLink}
+ *
+ * @deprecated this package has been created by mistake. The {@code LinkResult} moved to {@link
+ *     org.sdase.commons.server.jackson.hal.LinkResult}, please update the imports.
+ */
+@Deprecated
+public final class LinkResult extends org.sdase.commons.server.jackson.hal.LinkResult {
 
   /**
    * Instantiates a new Link result.
@@ -15,25 +18,6 @@ public final class LinkResult {
    * @param uri the uri
    */
   public LinkResult(URI uri) {
-    this.uri = uri;
-    this.halLink = new Builder(uri).build();
-  }
-
-  /**
-   * Returns the link result as {@linkplain URI}
-   *
-   * @return the uri
-   */
-  public URI asUri() {
-    return this.uri;
-  }
-
-  /**
-   * Returns the link result as {@linkplain HALLink}
-   *
-   * @return the hal link
-   */
-  public HALLink asHalLink() {
-    return halLink;
+    super(uri);
   }
 }
