+++ sda-commons-server-jackson/src/main/java/org/sdase/commons/server/jackson/JacksonConfigurationBundle.java
@@ -9,7 +9,6 @@ import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
 import java.time.ZonedDateTime;
 import java.util.function.Consumer;
-import org.sda.commons.server.jackson.hal.HalLinkProvider;
 import org.sdase.commons.server.jackson.errors.ApiExceptionMapper;
 import org.sdase.commons.server.jackson.errors.EarlyEofExceptionMapper;
 import org.sdase.commons.server.jackson.errors.JacksonPropertyNodeNameProvider;
@@ -19,6 +18,7 @@ import org.sdase.commons.server.jackson.errors.RuntimeExceptionMapper;
 import org.sdase.commons.server.jackson.errors.ValidationExceptionMapper;
 import org.sdase.commons.server.jackson.errors.WebApplicationExceptionMapper;
 import org.sdase.commons.server.jackson.filter.JacksonFieldFilterModule;
+import org.sdase.commons.server.jackson.hal.HalLinkProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
