+++ sda-commons-server-jackson/src/main/java/org/sdase/commons/server/jackson/hal/HalLinkMethodInvocationException.java
@@ -1,0 +1,12 @@
+package org.sdase.commons.server.jackson.hal;
+
+public class HalLinkMethodInvocationException extends RuntimeException {
+
+  public HalLinkMethodInvocationException(String message) {
+    super(message);
+  }
+
+  public HalLinkMethodInvocationException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}
