+++ sda-commons-server-jackson/src/test/java/org/sdase/commons/server/jackson/hal/HalLinkProviderTest.java
@@ -1,4 +1,4 @@
-package org.sda.commons.server.jackson.hal;
+package org.sdase.commons.server.jackson.hal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -12,37 +12,41 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.QueryParam;
 import org.junit.Test;
+import org.sda.commons.server.jackson.hal.HalLinkMethodInvocationException;
+import org.sda.commons.server.jackson.hal.LinkResult;
 
 public class HalLinkProviderTest {
 
   @Test
   public void shouldProvideHalLinkForNormalPathParams() {
-    final LinkResult linkResult = linkTo(methodOn(TestApi.class).testMethod("TEST"));
+    final org.sda.commons.server.jackson.hal.LinkResult linkResult =
+        linkTo(methodOn(TestApi.class).testMethod("TEST"));
     assertLinkResult(linkResult, "/testPath/TEST");
   }
 
   @Test
   public void shouldFailWhenNoInterfaceIsProvided() {
     assertThatThrownBy(() -> linkTo(methodOn(TestController.class).testMethod("FAIL")))
-        .isInstanceOf(HalLinkMethodInvocationException.class);
+        .isInstanceOf(org.sda.commons.server.jackson.hal.HalLinkMethodInvocationException.class);
   }
 
   @Test
   public void shouldProvideHalLinkForQueryParam() {
-    final LinkResult linkResult = linkTo(methodOn(TestApi.class).testMethodQueryParam("TEST"));
+    final org.sda.commons.server.jackson.hal.LinkResult linkResult =
+        linkTo(methodOn(TestApi.class).testMethodQueryParam("TEST"));
     assertLinkResult(linkResult, "/testPath?testRequestParam=TEST");
   }
 
   @Test
   public void shouldProvideHalLinkForDetailed() {
-    final LinkResult linkResult =
+    final org.sda.commons.server.jackson.hal.LinkResult linkResult =
         linkTo(methodOn(TestApi.class).testMethodDetail("TEST", 1, "testTheQuery"));
     assertLinkResult(linkResult, "/testPath/TEST/detail/testTheQuery?query=1");
   }
 
   @Test
   public void shouldProvideHalLinkForDetailedAndIgnoreQueryParamWithNullValue() {
-    final LinkResult linkResult =
+    final org.sda.commons.server.jackson.hal.LinkResult linkResult =
         linkTo(methodOn(TestApi.class).testMethodDetail("TEST", null, "testTheQuery"));
     assertLinkResult(linkResult, "/testPath/TEST/detail/testTheQuery");
   }
@@ -51,19 +55,20 @@ public class HalLinkProviderTest {
   public void shouldFailWithoutAnnotation() {
     assertThatThrownBy(
             () -> linkTo(methodOn(TestApi.class).testMethodWithoutPathParamAnnotation("FAIL")))
-        .isInstanceOf(HalLinkMethodInvocationException.class);
+        .isInstanceOf(org.sda.commons.server.jackson.hal.HalLinkMethodInvocationException.class);
   }
 
   @Test
   public void shouldDoNothingWhenNoParamsAreProvided() {
-    final LinkResult linkResult = linkTo(methodOn(TestApi.class).testMethodWithoutParams());
+    final org.sda.commons.server.jackson.hal.LinkResult linkResult =
+        linkTo(methodOn(TestApi.class).testMethodWithoutParams());
     assertLinkResult(linkResult, "/testPathWithNoParams");
   }
 
   @Test
   public void shouldFailWithNonProxiedMethod() {
     assertThatThrownBy(() -> linkTo("testMethod"))
-        .isInstanceOf(HalLinkMethodInvocationException.class);
+        .isInstanceOf(org.sda.commons.server.jackson.hal.HalLinkMethodInvocationException.class);
   }
 
   @Test
